removed the battery related functions, and variables, i.e. getBAtVolt , getChargeState(),
changed technique used to detect motor On/off,
only usign mFeedBack, line to detect motor on/of,
3 phase line is used to detect all 3 phases are present,
ACPhase line is used to detect if AC input to the unit is on/off,
the acFeedBack is used to detect whether any 2 phases are present or not, when the motor is off.

changed updateSensorData(), readSensorData(), anotherConstructor(),

added triggerAutoStart(). to detect and startTimer of turning motor on, if all the things are OK.
added motorState,ACPowerState,AllPhaseState,ACFeedBackState properties.

added checkLineSensors() to detect whether 3, 2 ,1 or no phase is present.

removed setACPowerState().
changed constructors signature.