added batStatus for last known status of the low battery pin  and batLevelChange to detect the change in low battery pin in ISR.
also, added reset reason detection in setup() which should only work when bootloader is not present ,as MCUSR is used in it.
also, init batStatus with status of the low battery pin.

changes ISR(PCINT0_vect) to detect whether there is a changes in low batt pin, or another pins. and report accordingly,

added turnoffTimer when low batt event is occured, to remove the false events of low batt.