changed anotherConstructor : to set waitCheckACTime to 2 secs. 								to set singlePhasingTime to 1 secs. changed readSensorData() : to not print ACFeedBack i.e. p3 changed updateSensorData(): to set the stop relay pin to low , starting  the relay, when AutoStart is ON.
changed resetAutoStart() : to stop the startTimerOn when setting the AutoStart to OFF.

changed operateOnEvent() : to turn led on or off when needed,
							to force stopMotor() when needed.

							in motor off previosly, than change motorState() to tMotor to check current state.

in checkLineSensors() : to turn led on or off, when needed.

changed unknownMotoroff() : to force stopMotor()
changed operateOnSinglePhasing() : to force StopMotor()

added setLED() : to turn LED on or off.
made checkLineSensors() public