Call Related , BYPASS, ADMIN NUMBER CHANGES
in Definitions.h : added admin Numbers

in Motor_MGR :
  changed bool allPhaseState() : added check for bypass to allow running even on single phasing

in SIM :
added isRegisteredNumber and set its value in needed functions

in operateOnMsg() : added 'RESPN' option in RESP section for no calls
in isAdmin() : added check for total 5 admin numbers, 2 personal and 3 company numbers
in operateRing() : added isRegisteredNumber to check if the number is registered or not for 3 rings, and cut if not registered after 3 rings. only accept if registeredNumber

in makeResponseAction() : added RESPONSE=='N' check, it is used to allow calls to be made for m2mEvents, to m2mRemote.
in registerEvent() : added check for RESPONSE == N, which allows to not make any call for regular events

in operateOnStagedEvent() : changed location of freezeIncomingCalls() and acceptCommands(), to allow accepting incoming calls even if event is staged, and avoid unnecessary wake up of sim.